requirements.txt
appnope==0.1.0
astroid==2.2.5
autopep8==1.4.4
backcall==0.1.0
bleach==3.0.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.1.0
django-heroku==0.3.1
django-pyodbc-azure==2.1.0.0
django-webpack-loader==0.6.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
entrypoints==0.2.3
gunicorn==19.9.0
idna==2.7
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.19
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.4.1
Markdown==3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pipenv==2018.11.26
prometheus-client==0.4.0
prompt-toolkit==2.0.5
psycopg2==2.7.6.1
psycopg2-binary==2.8.2
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.2.0
PyJWT==1.7.0
pylint==2.3.1
pyodbc==4.0.26
python-dateutil==2.7.3
pytz==2018.7
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.21.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
stripe==2.15.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.23
virtualenv==16.5.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.2
widgetsnbextension==3.4.2
wrapt==1.11.1
